Refactor of tests.
